RAG_requirement.txt
psycopg2-binary
sentence-transformers
ollama
google-generativeai
python-dotenv